fix: refatora componente de publicacao parcialmente
