+++ frontend/src/components/PublicationAntiga/styles.css
