+++ frontend/src/pages/home/index.js
@@ -2,6 +2,8 @@ import React from "react";
 
 import Header from "../../components/header";
 
+import Publication from '../../components/Publication';
+
 const Home = () => {
   return (
     <div className="container mx-auto max-w-[1580px]">
@@ -12,6 +14,7 @@ const Home = () => {
 
         {/* Design do componente de publication, fazer a seção do meio sem cor mesmo. Utilizar mais espaçamento da header para telas pequenas. Arrumar a pesquisa para filmes */}
         <div className="w-1/2 bg-white rounded-xl m-8 shadow-lg">
+          <Publication />
         </div>
 
         <div className="w-1/4"></div>
